cli/registry/client: deprecate and move internal
