+++ cli/command/manifest/client_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/distribution/reference"
 	manifesttypes "github.com/docker/cli/cli/manifest/types"
-	"github.com/docker/cli/cli/registry/client"
+	"github.com/docker/cli/internal/registryclient"
 	"github.com/docker/distribution"
 	"github.com/opencontainers/go-digest"
 )
@@ -45,4 +45,4 @@ func (c *fakeRegistryClient) PutManifest(ctx context.Context, ref reference.Name
 	return digest.Digest(""), nil
 }
 
-var _ client.RegistryClient = &fakeRegistryClient{}
+var _ registryclient.RegistryClient = &fakeRegistryClient{}
